requirements.txt
alembic==1.16.4
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-babel==4.0.0
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.3.0
python-barcode==0.15.1
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
gunicorn>=20.0.0
python-dotenv>=0.19.0
cryptography==41.0.0
Flask==2.3.2
Flask-WTF==1.0.0
gunicorn==20.1.0
python-dotenv==1.0.0